% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auto_select_best_kpclr_model}
\alias{auto_select_best_kpclr_model}
\title{Auto-Select Best KPCLR Model}
\usage{
auto_select_best_kpclr_model(explore_kpclr_obj, fp_max_cost = NULL,
  fn_min_cost = NULL, fp_min_cost = NULL, fn_max_cost = NULL)
}
\arguments{
\item{explore_kpcr_obj}{An object of type \code{explore_kpclr}.}

\item{fp_max_cost}{The maximum cost of a false positve. Together with \code{fn_min_cost}, this will inform
the algorithm of the maximum cost ratio of fp:fn. If left to the default \code{NULL}, the maximum
cost ratio will be 25\% more than the desired cost ratio.}

\item{fn_min_cost}{The minimum cost of a false negative. Together with \code{fp_max_cost}, this will inform
the algorithm of the maximum cost ratio of fp:fn. If left to the default \code{NULL}, the maximum
cost ratio will be 25\% more than the desired cost ratio.}

\item{fp_min_cost}{The minimum cost of a false positive. Together with \code{fn_max_cost}, this will inform
the algorithm of the minimum cost ratio of fp:fn. If left to the default \code{NULL}, the minimum
cost ratio will be 25\% less than the desired cost ratio.}

\item{fn_max_cost}{The maximum cost of a false negative. Together with \code{fp_min_cost}, this will inform
the algorithm of the minimum cost ratio of fp:fn. If left to the default \code{NULL}, the minimum
cost ratio will be 25\% less than the desired cost ratio.}
}
\value{
An object of type \code{explore_kpclr} with information regarding the auto-
								selected model.
}
\description{
Selects the best KPCLR model based based on finding the minimum cost-weighted error model that has a fn:fp ratio
within the bounds defined by \code{fp_max_cost}, \code{fn_min_cost}, \code{fp_min_cost}
and \code{fn_max_cost}.
}
\author{
Adam Kapelner and Justin Bleich
}

