% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auto_select_best_kpcr_model}
\alias{auto_select_best_kpcr_model}
\title{Auto-Select Best KPCR Model}
\usage{
auto_select_best_kpcr_model(explore_kpcr_obj)
}
\arguments{
\item{explore_kpcr_obj}{An object of type \code{explore_kpcr}.}
}
\value{
An object of type \code{explore_kpcr} with information regarding the auto-
								selected model.
}
\description{
Selects the best KPCR model based on finding the minimum sum of squared error on the validation data.
}
\author{
Adam Kapelner and Justin Bleich
}

