% Generated by roxygen2 (4.0.1): do not edit by hand
\name{best_kpclr_three_splits}
\alias{best_kpclr_three_splits}
\title{Best KPCLR with 3 Splits}
\usage{
best_kpclr_three_splits(X, y, kernel_list, split_props = c(1/3, 1/3, 1/3),
  rho_seq = seq(0.05, 0.95, 0.05), fn_to_fp_ratio = 1, plot = TRUE,
  tile_cols = 3)
}
\arguments{
\item{X}{The data's predictor matrix}

\item{y}{The data's response vector}

\item{kernel_list}{A list of kernels to assess performance of over a variety of PC's by \% explained.
Each element of this list is a list itself with keys "kernel_type" and "params."}

\item{split_props}{When splitting the data into training, validation and test sets respectively, what
proportions are assigned to each set? This must be numeric of size 3 and the numbers
will be normalized to sum to one. The default split is uniform (1/3, 1/3, 1/3).}

\item{rho_seq}{A collection of proportions of the variance explained of the kernel matrix to use}

\item{fn_to_fp_ratio}{The ratio of the "severity" of the false negative to the "severity" of the false positive
(defaults to 1)}
}
\value{
An object of class \code{best_kplcr} which is a list housing the results of the procedure
}
\description{
Performs the full procedure outlined in the paper. We first take three splits of the data
randomly and a cost ratio. Then, we take in a list of kernels. For each kernel, we build a
KPCR model on the training data (the first split). Then we look at performance of each kernel
over a range of # of principle components (by proportion of kernel matrix explained, \eqn{\rho}).
We pick the proportion that is closest to the predicted cost ratio for each kernel. Now, for
each kernel, we have a model. We then predict all models on the validation data (the second split)
and pick the best model. Using this best model, we predict on the test data (the third split)
and report the out-of-sample statistics
}
\author{
Justin Bleich and Adam Kapelner
}
\references{
Berk, R., Bleich, J., Kapelner, A.,  Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}
\seealso{
\code{\link{kpclr}}
}

