% Generated by roxygen2 (4.0.1): do not edit by hand
\name{build_final_kpclr_or_kpcr_model}
\alias{build_final_kpclr_or_kpcr_model}
\title{Create Final Kernel Model}
\usage{
build_final_kpclr_or_kpcr_model(explore_kpclr_or_kpcr)
}
\arguments{
\item{explore_kpclr_or_kpcr}{The object built from \code{explore_kplr_models} or \code{explore_kplcr_models}.}
}
\value{
The model corresponding to the \code{winning_kernel_num} and the
									\code{winning_rho_num} housed in the explore object.
}
\description{
Once the user has finished exploring different kernel regressions via \code{\link{explore_kplcr_models}} or
\code{\link{explore_kplr_models}} and has estimated future performance on the test data via
\code{\link{eval_winning_r_model_on_test_data}} or \code{\link{eval_winning_lr_model_on_test_data}},
we now build the final kernel model using all the data from \code{X}, \code{y}.
}
\author{
Adam Kapelner and Justin Bleich
}

