% Generated by roxygen2 (4.0.1): do not edit by hand
\name{build_kpca_object}
\alias{build_kpca_object}
\title{Builds a Kernel PCA Object}
\usage{
build_kpca_object(X, kernel_type, params = c())
}
\arguments{
\item{X}{The original design matrix}

\item{kernel_type}{One of the valid kernel types: vanilla, rbf, poly, tanh, bessel, laplace, anova, spline.}

\item{params}{A list of parameters specific to the kernel of the user's choice. Each kernel type has
a required number of parameters that must be passed otherwise the function will throw
an error.}
}
\value{
A list composed of the original data, the kernel, the matrix K, the non-zero eigenvalues and eigenvectors of K and K in the eigenbasis
}
\description{
Based on the original design matrix (the "data"), build an object which houses information about
the data in a transformed / kernelized space based on a kernel of the user's choice.
}
\author{
Justin Bleich and Adam Kapelner
}
\references{
Berk, R., Bleich, J., Kapelner, A.,  Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}
\seealso{
See \code{\link[kernlab]{dots}} for more information.
}

