% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kernel_finder_logistic_regression}
\alias{kernel_finder_logistic_regression}
\title{Searches Kernel Space}
\usage{
kernel_finder_logistic_regression(kernel_list, xtrain, ytrain, xtest, ytest,
  var_seq = seq(0.05, 0.95, by = 0.1), threshold = 0.5, weights = NULL,
  ...)
}
\arguments{
\item{kernel_list}{A list of kernels to assess performance of over a variety of PC's by \% explained.
Each element of this list is a list itself with keys "kernel_type" and "params."}

\item{xtrain}{The training data}

\item{ytrain}{The responses for the original training data}

\item{ytest}{The responses for the new testing data}

\item{xtest}{The features for the new testing data}

\item{var_seq}{A collection of proportions of the number of PC's to use when plotting performance}

\item{threshold}{The threshold of the probability estimate after which to call the estimate a "1"}

\item{weights}{A weight vector to use for the original training data}

\item{...}{Other parameters to be passed to the plot function}
}
\value{
A matrix of error rates for each proportion of the number of PC's in \code{var_seq}
}
\description{
Plots the performance in terms of the false negative and false positive error of a kernel PCA
logistic regression model at various proportions of principal components used over many provided
kernel models.
}
\author{
Justin Bleich and Adam Kapelner
}
\references{
Berk, R., Bleich, J., Kapelner, A.,  Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}
\seealso{
\code{\link{plot_kpca_logistic_regression_perf}}
}

