% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{kernReg}
\alias{kernReg}
\alias{kernReg-package}
\title{Linear Principal Components Kernel Regression Methods}
\description{
A tool to select and run kernel PCA linear models for regression and classification
}
\author{
Justin Bleich \email{bleich@wharton.upenn.edu} and Adam Kapelner \email{kapelner@wharton.upenn.edu}
}
\references{
Berk, R., Bleich, J., Kapelner, A.,  Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}
\keyword{Components}
\keyword{Kernel}
\keyword{Principal}
\keyword{Regression}
\keyword{Regression,}

