% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.explore_kpclr}
\alias{plot.explore_kpclr}
\title{Explore kpclr plot}
\usage{
\method{plot}{explore_kpclr}(x, ...)
}
\arguments{
\item{x}{The \code{explore_kpclr} object to plot}

\item{...}{Other parameters to pass to \code{\link{plot_explore_kpclr}}.}
}
\description{
This is an S3 convenience method for the function \code{\link{plot_explore_kpclr}}.
Please follow the link for the full documentation.
}
\author{
Adam Kapelner and Justin Bleich
}

