% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.explore_kpcr}
\alias{plot.explore_kpcr}
\title{Explore KPCR plot}
\usage{
\method{plot}{explore_kpcr}(x, ...)
}
\arguments{
\item{x}{The \code{explore_kpcr} object to plot}

\item{...}{Other parameters to pass to \code{\link{plot_explore_kpcr}}.}
}
\description{
This is an S3 convenience method for the function \code{\link{plot_explore_kpcr}}.
Please follow the link for the full documentation.
}
\author{
Adam Kapelner and Justin Bleich
}

