% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.kpca}
\alias{plot.kpca}
\title{Plots the kernel matrix}
\usage{
\method{plot}{kpca}(x, ...)
}
\arguments{
\item{x}{The kpca object to plot}

\item{...}{Other parameters to pass to \code{\link{plot_kpca}}.}
}
\description{
This is an S3 convenience method for the function \code{\link{plot_kpca}}.
Please follow the link for the full documentation.
}
\author{
Adam Kapelner and Justin Bleich
}

