% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_explore_kpclr}
\alias{plot_explore_kpclr}
\title{Explore kpclr plot}
\usage{
plot_explore_kpclr(explore_kpclr_obj, tile_cols = 2, ylim = NULL,
  min_fn_fp_ratio = NULL, max_fn_fp_ratio = NULL,
  quantile_aic_to_display = 0.75, quantile_cwe_to_display = 0.95,
  color_winning_model = "blue", color_num_fn_fp_ratio = "black",
  color_aic = "firebrick3", color_cwe = "forestgreen",
  show_rho_numbers = TRUE, text_label_offset_pct = 0.1,
  kernels_to_plot = NULL, ...)
}
\arguments{
\item{explore_kpclr_obj}{An object of type \code{explore_kpclr} built with \code{\link{explore_kpclr_models}}}

\item{tile_cols}{When plotting all kernel model performances, how many kernels per plot window column?
Default is \code{2}.}

\item{ylim}{The \code{ylim} parameter which is passed to the plot function.}

\item{min_fn_fp_ratio}{If specified, plots a horizontal line on the y-axis representing the lower bound of
the ratio of the number of false negatives to false positives. Defaults to the value set
by \code{\link{auto_select_best_kpclr_model}} (if it was run previously). If not, no line
is plotted.}

\item{max_fn_fp_ratio}{If specified, plots a horizontal line on the y-axis representing the upper bound of
the ratio of the number of false negatives to false positives. Defaults to the value set
by \code{\link{auto_select_best_kpclr_model}} (if it was run previously). If not, no line
is plotted.}

\item{quantile_aic_to_display}{When plotting the AICs for each model, which quantile should be truncated?
Default is \code{95\%}.}

\item{quantile_cwe_to_display}{When plotting the cost-weighted-errors for each model, which quantile should be
truncated? Default is \code{75\%}.}

\item{color_winning_model}{What color is the vertical line of the winning model. Default is blue.}

\item{color_num_fn_fp_ratio}{What color are the ratios of false negatives to false positive? The default is black.}

\item{color_aic}{What color are the AIC lines? Default is reddish.}

\item{color_cwe}{What color are the cost weighted error lines? Default is greenish.}

\item{show_rho_numbers}{Plot the rho number on each of the exploratory plots. Default is \code{TRUE}.}

\item{text_label_offset_pct}{If the rho numbers are plotted, what percent offset below the points? Default is \code{10\%}.}

\item{kernels_to_plot}{A list of indices of the kernels to plot. If left to the default \code{NULL}, all kernels are plotted.}

\item{...}{Other parameters to pass to plot. Of particular interest is \code{xlim} which will limit
                                        some models from being displayed.}
}
\description{
Many kernel principle components logistic regression models were fitted on the training data via
\code{\link{explore_kpclr_models}}. This function will create one plot for each of the kernel models investigated.
At every value of rho (the proportion of variance of the kernel matrix explained), the following three things
will be plotted (1) the ratio of the number of false negatives to the number of false positives when predicted
out of sample (2) the AIC of the model (3) the cost-weighted error of the out-of-sample validation data.
}
\examples{
\dontrun{
#pull the predictor matrix and dummify the response from the Boston Housing Data
data(Boston)
y = ifelse(Boston$medv > median(Boston$medv), 1, 0)
Boston$medv = NULL
X = as.matrix(Boston)
#now explore kernel models using the default kernel list and misclassification costs.
#Use parallelization for speed.
explore_kpclr_obj = explore_kpclr_models(X, y, num_cores = 4)
#now we plot to see how the models built on the training data performed on the validation data
plot_explore_kpclr(explore_kpclr_obj) #plot(explore_kpclr_obj) also works and is recommended
}
}
\author{
Adam Kapelner and Justin Bleich
}

