% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_kpca_logistic_regression_perf}
\alias{plot_kpca_logistic_regression_perf}
\title{Performance plot by \%PC's of a Kernel PCA Logistic Regression}
\usage{
plot_kpca_logistic_regression_perf(kpca_object, y_train, X_validate, y_validate,
  rho_seq = seq(0.05, 0.9, by = 0.05), threshold = 0.5, weights = NULL,
  legend = TRUE, ...)
}
\arguments{
\item{kpca_object}{The kernel PCA logistic regression model}

\item{y_train}{The responses for the original training data}

\item{X_validate}{The features for the new validation data}

\item{y_validate}{The responses for the new validation data}

\item{rho_seq}{A collection of proportions of the variance explained of the kernel matrix to use when plotting performance}

\item{threshold}{The threshold of the probability estimate after which to call the estimate a "1"}

\item{weights}{A weight vector to use for the original training data}

\item{legend}{Whether or not to show the legend (defaults to \code{TRUE})}

\item{...}{Other parameters to be passed to the plot function}
}
\value{
A matrix of error rates for each proportion of the number of PC's in \code{rho_seq}
}
\description{
Plots the performance in terms of the false negative and false positive error of a kernel PCA
logistic regression model at various proportions of principal components used. To plot the
performance, test data not seen during constrcution used to build the kernel PCA logistic
regression model is used.
}
\author{
Justin Bleich and Adam Kapelner
}
\references{
Berk, R., Bleich, J., Kapelner, A.,  Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}

