% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.kernReg}
\alias{predict.kernReg}
\title{Predicts for new data}
\usage{
\method{predict}{kernReg}(object, new_data, ...)
}
\arguments{
\item{object}{The Kernel PCA linear model object used to predict}

\item{new_data}{The new data the user wishes to predict}

\item{...}{Other parameters to be passed to \code{predict.lm}}
}
\value{
A vector of predictions with lenth of the number of rows of \code{new_data} generated via \code{predict.lm}
}
\description{
\code{predict.kernReg} predicts using the kernel PCA model for new data
}
\author{
Justin Bleich and Adam Kapelner
}
\seealso{
\code{predict.lm}
}

