% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.kpclr}
\alias{predict.kpclr}
\title{Predicts for new data}
\usage{
\method{predict}{kpclr}(object, new_data, type = "response", num_cores = 1,
  ...)
}
\arguments{
\item{object}{The Kernel PCA logistic model used to predict}

\item{new_data}{The new data the user wishes to predict}

\item{type}{Which output to return to the user. Use "response" for predicted probability and "link" for a logit (see \code{predict.glm} for more information)}

\item{num_cores}{Number of cores for parallel prediction}

\item{...}{Other parameters to be passed to \code{predict.glm}}
}
\value{
A vector of predictions with lenth of the number of rows of \code{new_data} generated via \code{predict.glm}
}
\description{
\code{predict.kpclr} predicts using the kernel PCA logistic model for new data
}
\examples{
\dontrun{
#first create binary classification data
X = matrix(rnorm(300), ncol = 4)
y = rbinom(300, 1, 0.5)
#build a KPCA object using the anova kernel with hyperparameters sigma = 0.1 and d = 3
kpca_obj = build_kpca_object(X, "anova", c(0.1, 3))
#build a kpclr model using 75\% of the variance in the kernel matrix and weights for 1:1 cost ratio
kpclr_mod = kpclr(kpca_obj, y, frac_var = 0.75, weights = weights_for_kpclr(y))
#create 10 new data records and forecast on the new data
x_star = matrix(rnorm(40), ncol = 4)
y_hat = predict(kpclr_mod, x_star)
}
}
\author{
Justin Bleich and Adam Kapelner
}
\seealso{
\code{\link{predict.glm}}
}

