% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.explore_kpclr}
\alias{print.explore_kpclr}
\title{Prints a summary of a explore_kpclr object}
\usage{
\method{print}{explore_kpclr}(x, ...)
}
\arguments{
\item{x}{The explore_kpclr object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a explore_kpclr object
}
\author{
Justin Bleich and Adam Kapelner
}

