% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.explore_kpcr}
\alias{print.explore_kpcr}
\title{Prints a summary of a explore_kpcr object}
\usage{
\method{print}{explore_kpcr}(x, ...)
}
\arguments{
\item{x}{The \code{explore_kpcr} object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a explore_kpcr object
}
\author{
Justin Bleich and Adam Kapelner
}

