% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.kernReg}
\alias{print.kernReg}
\title{Prints a summary of a kernReg object}
\usage{
\method{print}{kernReg}(x, ...)
}
\arguments{
\item{x}{The kernReg object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a kernReg object
}
\author{
Justin Bleich and Adam Kapelner
}

