% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.kpca}
\alias{print.kpca}
\title{Prints a summary of a kpca object}
\usage{
\method{print}{kpca}(x, ...)
}
\arguments{
\item{x}{The kpca object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a kpca object
}
\author{
Justin Bleich and Adam Kapelner
}

