% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.kpclr}
\alias{print.kpclr}
\title{Prints a summary of a kpclr object}
\usage{
\method{print}{kpclr}(x, ...)
}
\arguments{
\item{x}{The kpclr object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a kpclr object
}
\author{
Justin Bleich and Adam Kapelner
}

