% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.kpcr}
\alias{print.kpcr}
\title{Prints a summary of a kpcr object}
\usage{
\method{print}{kpcr}(x, ...)
}
\arguments{
\item{x}{The kpcr object to be summarized in the console}

\item{...}{Other parameters to pass to the default print function}
}
\description{
Prints a summary of a kpcr object
}
\author{
Justin Bleich and Adam Kapelner
}

