% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.explore_kpclr}
\alias{summary.explore_kpclr}
\title{Prints a summary of a explore_kpclr object}
\usage{
\method{summary}{explore_kpclr}(object, ...)
}
\arguments{
\item{object}{The explore_kpclr object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a explore_kpclr object
}
\author{
Justin Bleich and Adam Kapelner
}

