% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.explore_kpcr}
\alias{summary.explore_kpcr}
\title{Prints a summary of a explore_kpcr object}
\usage{
\method{summary}{explore_kpcr}(object, ...)
}
\arguments{
\item{object}{The \code{explore_kpcr} object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a explore_kpcr object
}
\author{
Justin Bleich and Adam Kapelner
}

