% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.kernLogReg}
\alias{summary.kernLogReg}
\title{Prints a summary of a kernLogReg object}
\usage{
\method{summary}{kernLogReg}(object, ...)
}
\arguments{
\item{object}{The kernLogReg object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a kernLogReg object
}
\author{
Justin Bleich and Adam Kapelner
}

