% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.kernReg}
\alias{summary.kernReg}
\title{Prints a summary of a kernReg object}
\usage{
\method{summary}{kernReg}(object, ...)
}
\arguments{
\item{object}{The kernReg object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a kernReg object
}
\author{
Justin Bleich and Adam Kapelner
}

