% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.kpca}
\alias{summary.kpca}
\title{Prints a summary of a kpca object}
\usage{
\method{summary}{kpca}(object, ...)
}
\arguments{
\item{object}{The kpca object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a kpca object
}
\author{
Justin Bleich and Adam Kapelner
}

