% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.kpclr}
\alias{summary.kpclr}
\title{Prints a summary of a kpclr object}
\usage{
\method{summary}{kpclr}(object, ...)
}
\arguments{
\item{object}{The kpclr object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a kpclr object
}
\author{
Justin Bleich and Adam Kapelner
}

