% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.kpcr}
\alias{summary.kpcr}
\title{Prints a summary of a kpcr object}
\usage{
\method{summary}{kpcr}(object, ...)
}
\arguments{
\item{object}{The kpcr object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a kpcr object
}
\author{
Justin Bleich and Adam Kapelner
}

