% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weights_for_kpca_logistic_regression}
\alias{weights_for_kpca_logistic_regression}
\title{Creates weights for kpca logistic regression}
\usage{
weights_for_kpca_logistic_regression(y_train, fn_to_fp_ratio = 1)
}
\arguments{
\item{y_train}{The responses for the training data - a binary vector of length \code{n}.}

\item{fn_to_fp_ratio}{The ratio of the "severity" of the false negative to the "severity" of the false positive (defaults to 1)}
}
\value{
A vector of weights of length \code{n}.
}
\description{
\code{weights_for_kpca_logistic_regression} will set individual weights
for each observation in the training data based on two things (1) the
ratio of 1's and 0's in the training set and (2) the user-set false-negative
to false-positive ratio.
}
\author{
Justin Bleich and Adam Kapelner
}

