% Generated by roxygen2 (4.0.1): do not edit by hand
\name{explore_kpclr_models}
\alias{explore_kpclr_models}
\title{Explore KPCLR for many kernels}
\usage{
explore_kpclr_models(X, y, kernel_list, seed = 0, split_props = c(1/3, 1/3,
  1/3), rho_seq = seq(0.05, 0.95, 0.05), fn_cost = 1, fp_cost = 1,
  fp_max_cost = NULL, fn_min_cost = NULL, fp_min_cost = NULL,
  fn_max_cost = NULL, num_cores = 1)
}
\arguments{
\item{X}{The data's predictor matrix.}

\item{y}{The data's response vector.}

\item{kernel_list}{A list of kernels to assess performance of over a variety of PC's by \% explained.
Each element of this list is a list itself with keys "kernel_type" and "params."}

\item{seed}{A seed to set the random generator. This is required because re-runs will result in
different training-validation-test splits which would allow a user to snoop on the
test data. We default this to 0 but please use your own seeds to avoid seed conflict.}

\item{split_props}{When splitting the data into training, validation and test sets respectively, what
proportions are assigned to each set? This must be numeric of size 3 and the numbers
will be normalized to sum to one. The default split is uniform (1/3, 1/3, 1/3).}

\item{rho_seq}{A collection of proportions of the variance explained of the kernel matrix to use.}

\item{fn_cost}{The target cost of a false negative (defaults to 1). Together with \code{fp_cost},
this will inform the algorithm of the desired ratio of costs.}

\item{fp_cost}{The target cost of a false positive (defaults to 1). Together with \code{fn_cost},
this will inform the algorithm of the desired cost ratio of fp:fn.}

\item{fp_max_cost}{The maximum cost of a false positve. Together with \code{fn_min_cost}, this will inform
the algorithm of the maximum cost ratio of fp:fn. If left to the default \code{NULL}, the maximum
cost ratio will be 25\% more than the desired cost ratio.}

\item{fn_min_cost}{The minimum cost of a false negative. Together with \code{fp_max_cost}, this will inform
the algorithm of the maximum cost ratio of fp:fn. If left to the default \code{NULL}, the maximum
cost ratio will be 25\% more than the desired cost ratio.}

\item{fp_min_cost}{The minimum cost of a false positive. Together with \code{fn_max_cost}, this will inform
the algorithm of the minimum cost ratio of fp:fn. If left to the default \code{NULL}, the minimum
cost ratio will be 25\% less than the desired cost ratio.}

\item{fn_max_cost}{The maximum cost of a false negative. Together with \code{fp_min_cost}, this will inform
                            the algorithm of the minimum cost ratio of fp:fn. If left to the default \code{NULL}, the minimum
                            cost ratio will be 25\% less than the desired cost ratio.}
}
\value{
An object of class \code{explore_kplcr} which is a list housing the results of the procedure
}
\description{
Performs the full procedure outlined in the paper. We first take three splits of the data
randomly and a cost ratio. Then, we take in a list of kernels. For each kernel, we build a
KPCR model on the training data (the first split). Then we look at performance of each kernel
over a range of # of principle components (by proportion of kernel matrix explained, \eqn{\rho}).
We pick the proportion that is closest to the predicted cost ratio for each kernel. Now, for
each kernel, we have a model. We then predict all models on the validation data (the second split)
and pick the best model. Using this best model, we predict on the test data (the third split)
and report the out-of-sample statistics
}
\author{
Justin Bleich and Adam Kapelner
}
\references{
Berk, R., Bleich, J., Kapelner, A., Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}
\seealso{
\code{\link{kpclr}}
}

