% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kpclr}
\alias{kpclr}
\title{Run a logistic regression using Kernel PCA}
\usage{
kpclr(kpca_object, y, num_pcs = NULL, frac_var = NULL, weights = NULL,
  family = "binomial")
}
\arguments{
\item{kpca_object}{The object that contains the kernel and the kernelized data with its eigendecomposition}

\item{y}{The response to be regressed on the features which are the principal components of the kernelized data}

\item{num_pcs}{The number of principal components to use for the regression (this or \code{frac_var} must be specified)}

\item{frac_var}{Pick the number of principal components to use based on the fraction of variance to explain (this or \code{num_pcs} must be specified)}

\item{weights}{Weights to be used on each observation in a weighted generalized least squares implementation. If not specified (default), uniform weights are used}

\item{family}{The family parameter to be passed to the glm function. Default is "binomial." Note that with this default, AIC calculations are not possible.}
}
\value{
An lm object with the kpca_object embedded as well as the number of principal components used
}
\description{
\code{kpclr} runs a logistic regression on features created from an eigendecomposition
of a certain dimension of the kernelized data
}
\examples{
\dontrun{
#pull the predictor matrix and dummify the response from the Boston Housing Data
data(Boston)
y = ifelse(Boston$medv > median(Boston$medv), 1, 0)
Boston$medv = NULL
X = as.matrix(Boston)
#build a KPCA object using the anova kernel with hyperparameters sigma = 0.1 and d = 3
kpca_obj = build_kpca_object(X, "anova", c(0.1, 3))
#build a kpclr model using 75\% of the variance in the kernel matrix and weights for 1:1 cost ratio
kpclr_mod = kpclr(kpca_obj, y, frac_var = 0.75, weights = weights_for_kpclr(y))
#print a summary of the model
kpclr_mod
}
}
\author{
Justin Bleich and Adam Kapelner
}
\references{
Berk, R., Bleich, J., Kapelner, A.,  Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}
\seealso{
\code{\link{kpcr}}
}

