% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.explore_kpclr}
\alias{plot.explore_kpclr}
\title{Explore kpclr plot}
\usage{
\method{plot}{explore_kpclr}(explore_kpclr_obj, plot_tile_cols = 3,
  quantile_aic_to_display = 0.75, quantile_cwe_to_display = 0.9,
  color_winning_model = "blue", color_num_fn_fp_ratio = "black",
  color_aic = "firebrick3", color_cwe = "forestgreen",
  show_rho_numbers = TRUE, text_label_offset_pct = 0.1, ...)
}
\arguments{
\item{explore_kpclr_obj}{An object of type \code{explore_kpclr} built with \code{\link{explore_kpclr_models}}}

\item{plot_tile_cols}{When plotting all kernel model performances, how many kernels per plot window column?
Default is \code{3}.}

\item{quantile_aic_to_display}{When plotting the AICs for each model, which quantile should be truncated?
Default is \code{90\%}.}

\item{quantile_cwe_to_display}{When plotting the cost-weighted-errors for each model, which quantile should be
truncated? Default is \code{75\%}.}

\item{color_winning_model}{What color is the vertical line of the winning model. Default is blue.}

\item{color_num_fn_fp_ratio}{What color are the ratios of false negatives to false positive? The default is black.}

\item{color_aic}{What color are the AIC lines? Default is reddish.}

\item{color_cwe}{What color are the cost weighted error lines? Default is greenish.}

\item{show_rho_numbers}{Plot the rho number on each of the exploratory plots. Default is \code{TRUE}.}

\item{text_label_offset_pct}{If the rho numbers are plotted, what percent offset below the points? Default is \code{10\%}.}

\item{...}{Other parameters to pass to plot.}
}
\description{
Many kernel principle components logistic regression models were fitted on the training data via
\code{\link{explore_kpclr_models}}. This function will create one plot for each of the kernel models investigated.
At every value of rho (the proportion of variance of the kernel matrix explained), the following three things
will be plotted (1) the ratio of the number of false negatives to the number of false positives when predicted
out of sample (2) the AIC of the model (3) the cost-weighted error of the out-of-sample validation data.
}
\author{
Adam Kapelner and Justin Bleich
}

