% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_explore_kpcr}
\alias{plot_explore_kpcr}
\title{Explore KPCR plot}
\usage{
plot_explore_kpcr(explore_kpcr_obj, tile_cols = 3,
  quantile_aic_to_display = 0.75, color_winning_model = "blue",
  color_sse = "black", color_aic = "firebrick3", show_rho_numbers = TRUE,
  text_label_offset_pct = 0.1, label_skip = 3, ...)
}
\arguments{
\item{explore_kpcr_obj}{An object of type \code{explore_kpcr} built with \code{\link{explore_kpcr_models}}}

\item{tile_cols}{When plotting all kernel model performances, how many kernels per plot window column?
Default is \code{3}.}

\item{quantile_aic_to_display}{When plotting the AICs for each model, which quantile should be truncated?
Default is \code{75\%}.}

\item{color_winning_model}{What color is the vertical line of the winning model. Default is blue.}

\item{color_aic}{What color are the AICs? Default is reddish.}

\item{color_sse}{What color are the SSEs? Default is greenish.}

\item{show_rho_numbers}{Plot the rho number on each of the exploratory plots. Default is \code{TRUE}.}

\item{text_label_offset_pct}{If the rho numbers are plotted, what percent offset below the points? Default is \code{10\%}.}

\item{label_skip}{If the rho numbers are plotted, how many should be skipped? Default is \code{3}.}

\item{...}{Other parameters to pass to plot.}
}
\description{
Many kernel principle components logistic regression models were fitted on the training data via
\code{\link{explore_kpcr_models}}. This function will create one plot for each of the kernel models investigated.
At every value of rho (the proportion of variance of the kernel matrix explained), the AIC of the model and the sse
of the out-of-sample validation data is plotted.
}
\author{
Adam Kapelner and Justin Bleich
}

