% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_pdp}
\alias{plot_pdp}
\title{Plots a PDP for a kernel regression}
\usage{
plot_pdp(kpca_model, predictor, type = "link", frac_to_build = 1, ...)
}
\arguments{
\item{kpca_model}{A linear or logistic kernel PCA regression model}

\item{predictor}{The predictor to build a PDP for. It can be the number of the column or the column's name}

\item{type}{The type argument that gets passed to the \code{kpca_predict} function}

\item{frac_to_build}{The fraction of the design matrix to construct the PDP from.}

\item{...}{Other parameters to pass to the \code{plot} function}
}
\description{
\code{plot_pdp} plots a partial dependence plot (PDP; Friedman, 2001) for one of the predictors. The function
allows the PDP to be plotted to arbitrary accuracy and allows parameters to be passed for customization
of the plotting. This function relies on package \pkg{ICEbox}.
}
\author{
Justin Bleich and Adam Kapelner
}
\references{
Berk, R., Bleich, J., Kapelner, A.,  Henderson, J. and Kurtz, E., Using Regression Kernels to Forecast A Failure to Appear in Court. (2014) working paper
}
\seealso{
\code{\link{predict.kernReg}}, \code{\link{predict.kernLogReg}}, \code{\link{plot}}
}

