% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.kpclr}
\alias{predict.kpclr}
\title{Predicts for new data}
\usage{
\method{predict}{kpclr}(object, new_data, type = "response", ...)
}
\arguments{
\item{object}{The Kernel PCA logistic model used to predict}

\item{new_data}{The new data the user wishes to predict}

\item{type}{Which output to return to the user. Use "response" for predicted probability and "link" for a logit (see \code{predict.glm} for more information)}

\item{...}{Other parameters to be passed to \code{predict.glm}}
}
\value{
A vector of predictions with lenth of the number of rows of \code{new_data} generated via \code{predict.glm}
}
\description{
\code{predict.kpclr} predicts using the kernel PCA logistic model for new data
}
\author{
Justin Bleich and Adam Kapelner
}
\seealso{
\code{predict.glm}
}

