% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_desired_model}
\alias{set_desired_model}
\title{Sets Desired Model}
\usage{
set_desired_model(explore_kpcr_or_kpclr_obj, winning_kernel_num,
  winning_rho_num)
}
\arguments{
\item{explore_kpcr_or_kpclr_obj}{The object created by running \code{\link{explore_kpcr_models}} or \code{\link{explore_kpclr_models}}}

\item{winning_kernel_num}{The desired model's kernel index.}

\item{winning_rho_num}{The desired model's rho index.}
}
\value{
An object of type \code{exlore_kpcr} or \code{explore_kpclr} augmented with
										information about the user's desired model.
}
\description{
Given an object created with \code{\link{explore_kpcr_models}} or \code{\link{explore_kpclr_models}} and plotted with
\code{\link{plot.explore_kpcr}} or \code{\link{plot.explore_kpclr}} respectively, the user now chooses a desired model
by selecting a kernel and a rho index from this visualization. This function simply inputs this information into the
object returned by the \code{\link{explore_kpclr_models}} or \code{\link{explore_kpclr_models}} functions. If the user
wishes to reset the desired model, run this function with \code{NULL} for the arguments \code{winning_kernel_num} and
\code{winning_rho_num}.
}
\author{
Adam Kapelner and Justin Bleich
}

